SAAS-236 Store check results instead of alerts
