+++ services/checks/registry_test.go
@@ -21,38 +21,63 @@ import (
 	"time"
 
 	"github.com/go-openapi/strfmt"
+	"github.com/percona-platform/saas/pkg/check"
 	"github.com/percona/pmm/api/alertmanager/ammodels"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestRegistry(t *testing.T) {
-	t.Run("Collect Alerts", func(t *testing.T) {
+	t.Run("Create and Collect Alerts", func(t *testing.T) {
 		r := newRegistry()
 
 		nowValue := time.Now().UTC().Round(0) // strip a monotonic clock reading
 		r.nowF = func() time.Time { return nowValue }
-
-		labels := map[string]string{"label": "demo"}
-		annotations := map[string]string{"annotation": "test"}
 		alertTTL := resolveTimeoutFactor * defaultResendInterval
-
-		expected := &ammodels.PostableAlert{
-			Annotations: annotations,
-			EndsAt:      strfmt.DateTime(nowValue.Add(alertTTL)),
-			Alert: ammodels.Alert{
-				Labels: labels,
+		checkResults := []sttCheckResult{
+			{
+				checkName: "name",
+				target: target{
+					agentID:   "/agent_id/123",
+					serviceID: "/service_id/123",
+					labels: map[string]string{
+						"foo": "bar",
+					},
+				},
+				result: check.Result{
+					Summary:     "check summary",
+					Description: "check description",
+					Severity:    check.Warning,
+					Labels: map[string]string{
+						"baz": "qux",
+					},
+				},
 			},
 		}
 
-		alerts := ammodels.PostableAlerts{
-			r.createAlert(labels, annotations, alertTTL),
-		}
-		r.set(alerts)
+		r.set(checkResults)
 
-		collectedAlerts := r.collect()
+		expectedAlert := &ammodels.PostableAlert{
+			Annotations: map[string]string{
+				"summary":     "check summary",
+				"description": "check description",
+			},
+			EndsAt: strfmt.DateTime(nowValue.Add(alertTTL)),
+			Alert: ammodels.Alert{
+				Labels: map[string]string{
+					"alert_id":  "/stt/3293a1002b32c425dbda851406bf1428dced74cd",
+					"alertname": "name",
+					"baz":       "qux",
+					"foo":       "bar",
+					"severity":  "warning",
+					"stt_check": "1",
+				},
+			},
+		}
+
+		collectedAlerts := r.collect(alertTTL)
 		require.Len(t, collectedAlerts, 1)
 		require.Equal(t, 1, cap(collectedAlerts))
-		assert.Equal(t, expected, collectedAlerts[0])
+		assert.Equal(t, expectedAlert, collectedAlerts[0])
 	})
 }
